feat(fizzbuzz): configured dependency injection
