+++ lib/injection_container.dart
@@ -1,0 +1,5 @@
+import 'package:get_it/get_it.dart';
+
+final sl = GetIt.instance;
+
+void init() {}
